==> backend/notebooks/3_player_career_stats.ipynb <==
import pandas as pd

df = pd.read_csv("../data/processed/player_match_summary.csv")
df.head()


bat_stats = df.groupby("player").agg(
    innings=("runs", "count"),
    runs=("runs", "sum"),
    balls=("balls", "sum"),
    fours=("fours", "sum"),
    sixes=("sixes", "sum"),
    dismissals=("dismissed", "sum")
).reset_index()

bat_stats["average"] = bat_stats.apply(
    lambda x: x["runs"] / x["dismissals"] if x["dismissals"] > 0 else x["runs"], axis=1
)

bat_stats["strike_rate"] = 100 * bat_stats["runs"] / bat_stats["balls"]
bat_stats["boundary_rate"] = (bat_stats["fours"] + bat_stats["sixes"]) / bat_stats["balls"]

bat_stats.fillna(0, inplace=True)
bat_stats.head(10)


bowl_stats = df.groupby("player").agg(
    balls_bowled=("balls_bowled", "sum"),
    runs_conceded=("runs_conceded", "sum"),
    wickets=("wicket", "sum")
).reset_index()

bowl_stats["overs"] = bowl_stats["balls_bowled"] // 6 + (bowl_stats["balls_bowled"] % 6) / 10

bowl_stats["economy"] = bowl_stats["runs_conceded"] / (bowl_stats["balls_bowled"] / 6)
bowl_stats["bowling_avg"] = bowl_stats["runs_conceded"] / bowl_stats["wickets"].replace(0, 1)
bowl_stats["bowling_strike_rate"] = bowl_stats["balls_bowled"] / bowl_stats["wickets"].replace(0, 1)

bowl_stats.fillna(0, inplace=True)
bowl_stats.head(10)


career_stats = pd.merge(bat_stats, bowl_stats, on="player", how="outer")
career_stats.fillna(0, inplace=True)
career_stats.head(20)


career_stats["batting_strength"] = (
    0.4 * career_stats["average"] +
    0.3 * career_stats["strike_rate"] +
    0.2 * career_stats["boundary_rate"] * 100 +
    0.1 * (career_stats["innings"] - career_stats["dismissals"])
)

career_stats["bowling_strength"] = (
    0.5 * career_stats["wickets"] -
    0.3 * career_stats["economy"] +
    0.2 * (1 / career_stats["bowling_strike_rate"].replace(0, 1))
)

career_stats["overall_strength"] = (
    career_stats["batting_strength"] + career_stats["bowling_strength"]
)


career_stats.to_csv("../data/processed/player_career_stats.csv", index=False)
print("Saved player_career_stats.csv!")


df_sorted = df.sort_values(["player", "match_id"])

# Last 5-match rolling averages using transform
df_sorted["recent_runs"] = (
    df_sorted.groupby("player")["runs"]
    .transform(lambda x: x.rolling(5, min_periods=1).mean())
)

df_sorted["recent_balls"] = (
    df_sorted.groupby("player")["balls"]
    .transform(lambda x: x.rolling(5, min_periods=1).mean())
)

df_sorted["recent_wickets"] = (
    df_sorted.groupby("player")["wicket"]
    .transform(lambda x: x.rolling(5, min_periods=1).mean())
)

# For economy: sum of last 5 balls / sum of last 5 overs
df_sorted["recent_runs_conceded"] = (
    df_sorted.groupby("player")["runs_conceded"]
    .transform(lambda x: x.rolling(5, min_periods=1).sum())
)

df_sorted["recent_balls_bowled"] = (
    df_sorted.groupby("player")["balls_bowled"]
    .transform(lambda x: x.rolling(5, min_periods=1).sum())
)

df_sorted["recent_economy"] = df_sorted["recent_runs_conceded"] / (
    df_sorted["recent_balls_bowled"] / 6
)

df_sorted["recent_economy"].replace([float('inf'), -float('inf')], 0, inplace=True)
df_sorted["recent_economy"].fillna(0, inplace=True)

df_sorted.head(20)


df_sorted.to_csv("../data/processed/player_recent_form.csv", index=False)
print("Saved player_recent_form.csv!")
